Rename integration to IronOS
