+++ homeassistant/components/iron_os/icons.json
