+++ homeassistant/components/iron_os/manifest.json
@@ -1,6 +1,6 @@
 {
-  "domain": "pinecil",
-  "name": "Pinecil",
+  "domain": "iron_os",
+  "name": "IronOS",
   "bluetooth": [
     {
       "service_uuid": "9eae1000-9d0d-48c5-aa55-33e27f9bc533",
@@ -10,7 +10,7 @@
   "codeowners": ["@tr4nt0r"],
   "config_flow": true,
   "dependencies": ["bluetooth_adapters"],
-  "documentation": "https://www.home-assistant.io/integrations/pinecil",
+  "documentation": "https://www.home-assistant.io/integrations/iron_os",
   "iot_class": "local_polling",
   "loggers": ["pynecil"],
   "requirements": ["pynecil==0.2.0"]
