+++ homeassistant/components/iron_os/strings.json
