+++ requirements_all.txt
@@ -2022,7 +2022,7 @@ pymsteams==0.1.12
 # homeassistant.components.mysensors
 pymysensors==0.24.0
 
-# homeassistant.components.pinecil
+# homeassistant.components.iron_os
 pynecil==0.2.0
 
 # homeassistant.components.netgear
