+++ requirements_test_all.txt
@@ -1603,7 +1603,7 @@ pymonoprice==0.4
 # homeassistant.components.mysensors
 pymysensors==0.24.0
 
-# homeassistant.components.pinecil
+# homeassistant.components.iron_os
 pynecil==0.2.0
 
 # homeassistant.components.netgear
